Moving config to separate package
